requirements.txt
pandas
psycopg2-binary
requests
apache-airflow==2.9.1
